LayoutTests/platform/wpe/storage/domstorage/events/case-sensitive-expected.txt

Verify that storage events fire even when only the case of the value changes.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Testing sessionStorage
storage.clear()
PASS storage.length is 0

Verify storage events are case sensitive
storage.foo = 'test'
Reset storage event list
storageEventList = new Array()
storage.foo = 'test'
PASS storageEventList.length is 0
storage.foo = 'TEST'
PASS storageEventList.length is 1


Testing localStorage
storage.clear()
PASS storage.length is 0

Verify storage events are case sensitive
storage.foo = 'test'
FAIL storage.foo = 'test' threw exception QuotaExceededError: The quota has been exceeded.
FAIL Timeout: only got 0, expected 1 events
Reset storage event list
storageEventList = new Array()
storage.foo = 'test'
FAIL storage.foo = 'test' threw exception QuotaExceededError: The quota has been exceeded.
PASS storageEventList.length is 0
storage.foo = 'TEST'
FAIL storage.foo = 'TEST' threw exception QuotaExceededError: The quota has been exceeded.
FAIL Timeout: only got 0, expected 1 events
FAIL storageEventList.length should be 1. Was 0.


PASS successfullyParsed is true

TEST COMPLETE
